refactor: move resolve_and_store from commands/add to core/resolver
